refactor(base-app): base app parts
