+++ apps/base-app/src/app/ui.ts
@@ -1,0 +1,1 @@
+export { mountApp } from '@/base/ui'
