+++ apps/base-app/src/base/ui/providers/withVueConfig.ts
