+++ apps/base-app/src/base/ui/providers/withi18n/fr.json
